# Optimize stringify item before pushing it