@dsantosmerino 2020-05-22
> I believe there are several copies of this in the codebase. Can you find any more instances?

Yes, I found two more, one in the other way, symbolizing. Already changed both, let me know if you anything else to consider.